Pepper: Further fixes for the main interaction between Pepper and User
//TODO:
(1) - "Change favorite game to Semantic Ping Pong" button doesn't redirect from TicTacToe towards SemanticPP
(2) - Implement pepperSpeak() for welcoming the user in TicTacToe
(3) - Disable Send button when Pepper is thinking about a word in Semantic Ping Pong
(4) - Lower a bit Pepper thinking delay in SPP
(5) - We should have a "loading screen" at the beginning. When Pepper recognizes the user, it automatically redirects him towards ":5001/". This should be the case even when the user leaves the Pepper interaction (we must not stay in the SPP or TicTacToe screen, but back to the loading screen for another cycle of identification)
(6) - Fix Pepper losing dance
(7) - During the second iteration of identification of user in Pepper/main.py, the Master DB crashes

Signed-off-by: Gianmarco Scarano <gianmarcoscarano@gmail.com>